refactor: Export core-database typedefs (#1906)
