+++ packages/core-api/src/repositories/repository.ts
@@ -1,12 +1,13 @@
 import { app } from "@arkecosystem/core-container";
 import { PostgresConnection } from "@arkecosystem/core-database-postgres";
+import { TransactionPool } from "@arkecosystem/core-transaction-pool";
 import snakeCase from "lodash/snakeCase";
 import { IRepository } from "../interfaces/repository";
 
 export abstract class Repository implements IRepository {
     public database = app.resolvePlugin<PostgresConnection>("database");
     public cache = this.database.getCache();
-    public transactionPool = app.resolvePlugin("transactionPool");
+    public transactionPool = app.resolvePlugin<TransactionPool>("transactionPool");
     public model = this.getModel();
     public query = this.model.query();
     public columns: string[] = [];
