+++ packages/core-blockchain/package.json
@@ -33,6 +33,7 @@
         "@arkecosystem/core-database-postgres": "^2.1.0",
         "@arkecosystem/core-logger": "^2.1.0",
         "@arkecosystem/core-utils": "^2.1.0",
+        "@arkecosystem/core-transaction-pool": "^2.1.0",
         "@arkecosystem/crypto": "^2.1.0",
         "@types/lodash.get": "^4.4.4",
         "async": "^2.6.1",
