+++ packages/core-graphql/package.json
@@ -31,6 +31,7 @@
         "@arkecosystem/core-database-postgres": "^2.1.0",
         "@arkecosystem/core-logger": "^2.1.0",
         "@arkecosystem/core-http-utils": "^2.1.0",
+        "@arkecosystem/core-transaction-pool": "^2.1.0",
         "@arkecosystem/crypto": "^2.1.0",
         "apollo-server-hapi": "^2.3.1",
         "dayjs-ext": "^2.2.0",
