+++ packages/core-transaction-pool/src/guard.ts
@@ -3,12 +3,12 @@ import { PostgresConnection } from "@arkecosystem/core-database-postgres";
 import { AbstractLogger } from "@arkecosystem/core-logger";
 import { configManager, constants, models, slots } from "@arkecosystem/crypto";
 import pluralize from "pluralize";
+import { TransactionPool } from "./connection";
+import { dynamicFeeMatcher } from "./dynamic-fee";
+import { isRecipientOnActiveNetwork } from "./utils/is-on-active-network";
 
 const { TransactionTypes } = constants;
 const { Transaction } = models;
-
-import { dynamicFeeMatcher } from "./dynamic-fee";
-import { isRecipientOnActiveNetwork } from "./utils/is-on-active-network";
 
 export class TransactionGuard {
     public transactions: any[];
@@ -17,14 +17,14 @@ export class TransactionGuard {
     public broadcast: { [key: string]: any };
     public invalid: { [key: string]: any };
     public errors: any;
-    private pool: any;
+    private pool: TransactionPool;
 
     /**
      * Create a new transaction guard instance.
      * @param  {TransactionPoolInterface} pool
      * @return {void}
      */
-    constructor(pool) {
+    constructor(pool: TransactionPool) {
         this.pool = pool;
 
         this.transactions = [];
